backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
httpx==0.27.0
mistralai==0.4.2
openai-whisper==20231117
pydub==0.25.1
pytesseract==0.3.10
Pillow==10.3.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
aiofiles==23.2.1
gtts==2.5.1
